Code refactor to use BaseReport when generating console and HTMl reports.
